Change with a minimum number of coins
Construct a greedy algorithm that receives as input a natural number n and in time O(n) finds the minimum number of coins of denominations 1 penny, 5 penny, 10 penny and 25 penny that can be used to give change of n n n pennies. (As always, you need to describe the algorithm, prove its correctness, and evaluate it for the running time.